requirements.txt
appdirs==1.4.3
arabic-reshaper==2.1.3
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.1
blinker==1.4
Brlapi==0.7.0
caffeine==2.9.8
certifi==2019.11.28
chardet==3.0.4
cliapp==1.20180812.1
Click==7.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.7
django-ckeditor==6.1.0
django-cors-headers==3.11.0
django-countries==7.2.1
django-crontab==0.7.1
django-currentuser==0.5.3
django-dbbackup==3.3.0
django-debug-toolbar==3.2.2
django-heroku==0.3.1
django-import-export==2.6.0
django-js-asset==1.2.2
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.3
et-xmlfile==1.1.0
ewmh==0.1.6
filelock==3.4.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
Mako==1.1.0
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==2.1.0
meson==0.53.2
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
odfpy==1.4.1
onboard==1.4.1
openpyxl==3.0.9
packaging==20.3
Pillow==8.3.2
pipenv==2021.11.23
platformdirs==2.4.0
psycopg2==2.9.1
pycairo==1.16.2
pycares==2.4.0
pycodestyle==2.7.0
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
PyQt5==5.14.1
python-apt===2.0.0-ubuntu0.20.04.7-elementary7-ubuntu6.1.1
python-barcode==0.13.1
python-bidi==0.4.2
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-xlib==0.23
pytz==2021.1
pyxdg==0.26
PyYAML==5.4.1
reportlab==3.6.1
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.10
tablib==3.0.0
toml==0.10.2
ttystatus==0.38
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
virtualenv==20.10.0
virtualenv-clone==0.5.7
wadllib==1.3.3
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
